// lib/Type.ts
export interface SolidIconType {
  id: number;
  icon: React.ReactElement;
  name: string;
}
export interface RegularIconType {
  id: number;
  icon: React.ReactElement;
  link: string;
}
export interface ColorIconType {
  id: number;
  icon: React.ReactElement;
  link: string;
  title: string;
}
export interface ShortCutType {
  id: number;
  image: string;
  link: string;
  title: string;
  members: string;
  type: string;
}